Added Base modules for login and onetime_pass so that they can be more easily customized
